+++ CHANGES.md
@@ -1,5 +1,5 @@
 # Version 2.0.2
-Unreleased
+January 10, 2023
 
 - Implement reflection for array types (#213)
 - Fix get_multi_columns() to support multiple table names in filter_array (#220)